refactor : index.ts 를 활용해서 import문 수정
