+++ src/components/cartSummary/index.ts
@@ -1,0 +1,2 @@
+export { CartSummary } from "./cartSummary/CartSummary";
+export { CartSummaryItem } from "./cartSummaryItem/CartSummaryItem";
