+++ src/pages/cartPage/CartPage.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useRecoilState, useRecoilValue } from "recoil";
 import { getCartItems } from "../../api";
-import { ConfirmButton } from "../../components/button/confirmButton/ConfirmButton";
+import { ConfirmButton } from "../../components/button";
 import { CartContentSection } from "../../components/cartContentSection/CartContentSection";
 import { CartHeader } from "../../components/cartHeader/CartHeader";
 import Header from "../../components/header/Header";
